Move notify listener tests to test_core (#591)
